bot/requirements.txt
python-telegram-bot==20.7
paramiko>=2.12.0
python-dotenv>=1.0.0


